Add test document

+++ koztumize/tests/__init__.py
@@ -4,9 +4,11 @@ import os
 import shutil
 import pynuts
 from tempfile import mkdtemp
+from brigit import Git
 from koztumize import application
 
 PATH = os.path.dirname(os.path.dirname(__file__))
+TEMP_DIR = None
 
 
 def execute_sql(app, filename, folder=None):
@@ -22,6 +24,10 @@ def execute_sql(app, filename, folder=None):
 
 
 def setup():
+    # global variable shouldn't be used but is quite useful here
+    # pylint: disable=W0603
+    global TEMP_DIR
+    TEMP_DIR = mkdtemp()
     if not os.path.exists(
         os.path.join(PATH, 'tests', 'fake_instance', 'documents.git')):
         shutil.copytree(
@@ -30,12 +36,21 @@ def setup():
     config_file = os.path.join(PATH, 'config', 'test.cfg')
     app = application.Koztumize('koztumize', config_file=config_file)
     application.app = app
+    app.config['MODELS'] = os.path.join(TEMP_DIR, 'models')
+    git = Git(app.config['MODELS'])
+    git.init()
+    git.remote('add', '-t', 'models', 'origin', app.config['GIT_REMOTE'])
+    git.pull()
+    git.checkout('models')
     execute_sql(app, 'db.sql', 'tests')
     from koztumize import routes
+    import document
     
 
 def teardown():
     """Remove the temp directory after the tests."""
     execute_sql(application.app, 'drop_all.sql', 'tests')
-    shutil.rmtree(os.path.join(PATH, 'tests', 'fake_instance'))
-
+    if os.path.exists(os.path.join(PATH, 'tests', 'fake_instance')):
+        shutil.rmtree(os.path.join(PATH, 'tests', 'fake_instance'))
+    if os.path.exists(TEMP_DIR):
+        shutil.rmtree(TEMP_DIR)
